requirements.txt
pyautogui
Pillow
numpy
pygetwindow